requirements.txt
crewai==0.22.6
google-api-python-client==2.126.0
requests==2.31.0
python-dotenv==1.0.1
together==0.2.10
